+++ test/debug-mode.test.js
@@ -11,7 +11,7 @@ var engine = simplet({
 tap.test('result check', function (test) {
 
 	engine.template('t1', '<% print(\'simplet\') %>');
-	test.ok(engine.container['t1.ejs'].compiled === 'var _l_=1;try{print(\'simplet\')}catch(e){throw{line:_l_,message:e.message,name:e.name}}');
+	test.ok(engine.container['t1.ejs'].compiled === 'var _l_=0;try{print(\'simplet\')}catch(e){e.line=_l_;throw e}');
 
 	test.end();
 });